+++ lib/openapi_client/models/page_payment_list_response_dto.rb
@@ -38,7 +38,7 @@ module OpenapiClient
     # Attribute type mapping.
     def self.openapi_types
       {
-        :'total_row_count' => :'Float',
+        :'total_row_count' => :'Integer',
         :'total_amount' => :'Float',
         :'data' => :'Array<PagePaymentListResponseDTODataInner>'
       }
