+++ lib/openapi_client/models/payment_link_response_callback_parameters.rb
@@ -68,7 +68,7 @@ module OpenapiClient
         :'transaction_id' => :'String',
         :'callback_api_url' => :'String',
         :'callback_message' => :'String',
-        :'redirection_time' => :'Float'
+        :'redirection_time' => :'Integer'
       }
     end
