+++ lib/openapi_client/models/pf_generate_quote_response_item_e_sign_result.rb
@@ -58,7 +58,7 @@ module OpenapiClient
         :'batch_id' => :'String',
         :'agent_url' => :'String',
         :'insured_url' => :'String',
-        :'e_sign_option' => :'Float'
+        :'e_sign_option' => :'Integer'
       }
     end
