+++ lib/openapi_client/models/pf_lite_payment_link_request_callback_parameters.rb
@@ -60,7 +60,7 @@ module OpenapiClient
         :'reference_no' => :'String',
         :'reference_type' => :'String',
         :'transaction_id' => :'String',
-        :'redirection_time' => :'Float'
+        :'redirection_time' => :'Integer'
       }
     end
